digimitraai/requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.2
faiss-cpu>=1.7.4
openai>=1.3.0
streamlit>=1.29.0
python-dotenv>=1.0.0
openai-whisper
pyyaml>=6.0.1
numpy>=1.24.0
pandas>=2.0.0
tiktoken>=0.5.1
langchain-community
python-dotenv
sounddevice>=0.4.6
wavio>=0.0.7
PyAudio>=0.2.13
scipy
scikit-learn
PyPDF2
google-cloud-speech>=2.6.0
google-cloud-texttospeech>=2.12.0
google-cloud-translate>=2.0.1
soundfile
openai-whisper>=20231117
gradio>=4.4.0
pymupdf